+++ src/test/java/uk/gov/hmcts/opal/service/proxy/DefendantAccountEnforcementServiceProxyTest.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.opal.service.proxy;
 
 import tools.jackson.core.JacksonException;
+import java.math.BigInteger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -97,7 +98,7 @@ class DefendantAccountEnforcementServiceProxyTest extends ProxyTestsBase {
         AddEnforcementResponse expected = AddEnforcementResponse.builder()
             .enforcementId("ENF-L")
             .defendantAccountId("77")
-            .version(3)
+            .version(BigInteger.valueOf(3L))
             .build();
 
         when(legacyService.addEnforcement(defendantAccountId, businessUnitId,
@@ -131,7 +132,7 @@ class DefendantAccountEnforcementServiceProxyTest extends ProxyTestsBase {
         AddEnforcementResponse expected = AddEnforcementResponse.builder()
             .enforcementId("ENF-O")
             .defendantAccountId("77")
-            .version(3)
+            .version(BigInteger.valueOf(3L))
             .build();
 
         when(opalService.addEnforcement(defendantAccountId, businessUnitId,
